requirements.txt
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
requests==2.28.1
selenium==4.4.3
SQLAlchemy==1.4.41
black==22.8.0